11/23/2022: - Fixed download issue & added other improvements
Meta class is now in kotlin to fix a compilation error.

Added a couple extra checks for free space check to make it more reliable.

Added a new option to bypass the free space check.
- This check is not very consistent and is based on a lot of permissions and security checks.
- If you are having an issue, you can use this option.

Now when downloading with an episode url it no longer checks for identity.
- Still checks when finding the series for the episode, but fixes an issue

Added another clearer message when trying to use a browser you don't have installed.
- This can also mean you have folder permission issues.

Now when downloading a video that already exists in the download table, it re-adds it with a new date to push it to the top of the list.

Fixed the download issue.
- Cleaned up VideoDownloader to run more effectively
- Cannot have anymore memory leaks due to errors.
- A bit faster

Fixed a couple shutdown errors.