12/4/2022: - Links upgrade
Links are now handled better and stored in their own box instead of a class full of lists.
- CategoryLink now stores them with their identity type
- Faster and easier to manage
- Less prone to errors
- Stored in its own database folder for distribution

Now if the links are actually empty, it will download missing ones.

Now if a series could not be identified, it looks online and caches the link it matches
instead of downloading the entire list.

Added a new identity as NEW.
- This fixes a lot of issues where some series kept getting marked as NONE and skipped.
- This identity occurs when wco has not added a series to their list.

Created a new dynamic multi choice window for certain options.

WcoController now gives you the multi choice option when clicking on a series genre.

WcoController now gives you a multi choice option to label an uncategorized series as any identity.