12/6/2022: - Final touches
WcoControllers type updater can now be used all the time.

Added an Update Details option to WcoController in case a series is missing something.

Added the Update Details option to Series Details as well

Added a way to export debug errors to a file.

Now inside VideoDownloader if the frame couldn't be found, it exports the page source to a file.
- It will be in the debug folder. debug/source.txt
- This will be used so i can see how the web page has loaded when someone reports an error.