Updated version to 2.0
Added an option to change the domain and domain extension of the website to scrape.
- This is used in case the website changes entirely and must be manually updated.
- The default domain is wcofun and the default domain extension is org. (for now)
- Has an Auto Update button just in case.

Fixed a couple of unspecific issues and removed some unused code.

Added an updated useragents file to the github repository.
- The program now uses this list instead of my ancient one from years ago.

DownloadConfirm and SeriesDetails now no longer forces the width of the image to be the entire window.
All these changes are to allow a better UX.
- adjusted the image height to be smaller (no longer forced to be 1/3 of the window)
- deleted the "Description" text

Added a new option to download videos in higher supported resolutions.
- The default is 576p
- The 3 options are 576p, 720p and 1080p
- Not all videos support different resolutions.
- Changing the resolution also adds a (720p) or (1080p) at the end of the video file name and download name.
= supports downloading the same video with different resolutions
- If 1080 is selected, but the video only supports 720 then it will use 720
- works just as fast as the traditional way.
- If the resolution process fails, it falls back to the traditional way.

Toast now no longer changes window focus and won't bring itself to the front if you are using
another program.

Fixed series details window genre and episode loading when
updating series details

All video related objects now contain and use slugs.
- all link variables are deprecated.
- all slugs update on older data for the first launch
- these slugs will be used with the new video url option. This is so all
objects pointing to wco will now just use the video url option since the slugs never change, but the domain does.

Now when downloading with an individual episode link, it looks for cached series.

Fixed the option to check all series history for new episodes in Settings.

Officially made the new database system compatible with all old data.

Remade the wco db and compressed it to be 30MB for the new system.

Disabled the option to update the wco db.